Moved signing functions used from actor to builder/sighash.rs
