+++ core/src/actor.rs
@@ -287,36 +287,6 @@ impl Actor {
     }
 
     #[tracing::instrument(err(level = tracing::Level::ERROR), ret(level = tracing::Level::TRACE))]
-    pub fn convert_tx_to_script_spend(
-        tx_handler: &mut TxHandler,
-        txin_index: usize,
-        script_index: usize,
-        sighash_type: Option<TapSighashType>,
-    ) -> Result<TapSighash, BridgeError> {
-        let mut sighash_cache: SighashCache<&mut bitcoin::Transaction> =
-            SighashCache::new(&mut tx_handler.tx);
-
-        let prevouts = bitcoin::sighash::Prevouts::All(&tx_handler.prevouts);
-        let leaf_hash = TapLeafHash::from_script(
-            tx_handler
-                .scripts
-                .get(txin_index)
-                .ok_or(BridgeError::NoScriptsForTxIn(txin_index))?
-                .get(script_index)
-                .ok_or(BridgeError::NoScriptAtIndex(script_index))?,
-            LeafVersion::TapScript,
-        );
-        let sig_hash = sighash_cache.taproot_script_spend_signature_hash(
-            txin_index,
-            &prevouts,
-            leaf_hash,
-            sighash_type.unwrap_or(TapSighashType::Default),
-        )?;
-
-        Ok(sig_hash)
-    }
-
-    #[tracing::instrument(err(level = tracing::Level::ERROR), ret(level = tracing::Level::TRACE))]
     pub fn convert_tx_to_sighash_pubkey_spend(
         tx: &mut TxHandler,
         txin_index: usize,
@@ -328,24 +298,6 @@ impl Actor {
             txin_index,
             &bitcoin::sighash::Prevouts::All(&tx.prevouts),
             bitcoin::sighash::TapSighashType::Default,
-        )?;
-
-        Ok(sig_hash)
-    }
-
-    #[tracing::instrument(err(level = tracing::Level::ERROR), ret(level = tracing::Level::TRACE))]
-    pub fn convert_tx_to_pubkey_spend(
-        tx: &mut TxHandler,
-        txin_index: usize,
-        sighash_type: Option<TapSighashType>,
-    ) -> Result<TapSighash, BridgeError> {
-        let mut sighash_cache: SighashCache<&mut bitcoin::Transaction> =
-            SighashCache::new(&mut tx.tx);
-
-        let sig_hash = sighash_cache.taproot_key_spend_signature_hash(
-            txin_index,
-            &bitcoin::sighash::Prevouts::All(&tx.prevouts),
-            sighash_type.unwrap_or(TapSighashType::Default),
         )?;
 
         Ok(sig_hash)
